requirements.txt
streamlit
requests
beautifulsoup4
python-dotenv
PyPDF2
gspread
google-auth
google-auth-oauthlib
sentence-transformers
fpdf2